Move serialization related classes to Serialization namespace (#3)
