--- src/StateMachine/ActionFactory.php
@@ -1,9 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace FireGento\MageBot\StateMachine;
-
-interface ActionFactory
-{
-    public function create(string $type, array $parameters);
-}
