+++ src/StateMachine/Serialization/NewActionFactory.php
@@ -1,7 +1,9 @@
 <?php
 declare(strict_types=1);
 
-namespace FireGento\MageBot\StateMachine;
+namespace FireGento\MageBot\StateMachine\Serialization;
+
+use FireGento\MageBot\StateMachine\Serialization\ActionFactory;
 
 /**
  * Minimal implementation of action factory. Assumes, $type is a class name and $parameters a single constructor argument
